# src/main/java/com/juegocartas/juegocartas/controller/rest/AuthController.java
package com.juegocartas.juegocartas.controller.rest;

import com.juegocartas.juegocartas.dto.request.LoginRequest;
import com.juegocartas.juegocartas.dto.request.RegisterRequest;
import com.juegocartas.juegocartas.dto.response.AuthResponse;
import com.juegocartas.juegocartas.service.AuthService;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/auth")
public class AuthController {

    private final AuthService authService;

    public AuthController(AuthService authService) {
        this.authService = authService;
    }

    @PostMapping("/register")
    public ResponseEntity<AuthResponse> register(@Valid @RequestBody RegisterRequest request) {
        return ResponseEntity.ok(authService.register(request));
    }

    @PostMapping("/login")
    public ResponseEntity<AuthResponse> login(@Valid @RequestBody LoginRequest request) {
        return ResponseEntity.ok(authService.login(request));
    }
}